+++ NEWS.md
@@ -1,7 +1,17 @@
+# document 2.0.0
+
+* document() now throws an error if the temporary package does not pass R CMD
+  check without errors, warnings or notes. 
+  - Added argument 'stop\_on\_check\_not\_passing'. Set to FALSE to issue a
+    warning() instead of throwing an error.
+  - Added argument 'check\_as\_cran' to use the --as-cran flag with R CMD check,
+    document() then assumes a single NOTE (on the CRAN incoming feasibility) to
+    be okay.
+
 # document 1.2.1
 
 * Check for RStudio in man().
-* Fix return status of write_the_docs().
+* Fix return status of write\_the\_docs().
 
 # document 1.2.0